requirements.txt
mediapipe
opencv-python-headless
streamlit-webrtc
numpy
av
streamlit
Pillow